Version Testing 2.0 Input/Output works now. A button has been added to mainactivity for testing since previously the activity would automatically test petapp but this caused a buffer over load. So i have it so the petapp only starts when the button is clicked.
A json object is now returned when the petApp closes, it returns the data inputed from the forum.
A json object is also sent to petApp to signal weather it is New or Edit version.

To do:
Setup methods to utilize the incoming data.
Currently there it no Edit pet version, only new.